mashora-site/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.0
celery[redis]==5.4.0
redis==5.1.0
miniopy-async==1.21.0
stripe==10.0.0
emails==0.6
jinja2==3.1.4
python-dotenv==1.0.1
